patched for 64bit support

+++ sdlBasic/gtk/SciTEGTK.cxx
@@ -2219,7 +2219,7 @@ gint SciTEGTK::QuitSignal(GtkWidget *, GdkEventAny *, SciTEGTK *scitew) {
 }
 
 void SciTEGTK::ButtonSignal(GtkWidget *, gpointer data) {
-	instance->Command((guint)data);
+	instance->Command((intptr_t)data);
 }
 
 void SciTEGTK::MenuSignal(SciTEGTK *scitew, guint action, GtkWidget *) {
